create kasus_reservation

+++ database/migrations/2022_10_08_102519_create_pelapor_file.php
@@ -15,7 +15,7 @@ class CreatePelaporFile extends Migration
     {
         Schema::create('pelapor_file', function (Blueprint $table) {
             $table->id();
-            $table->integer('id_kasus');
+            $table->unsignedBigInteger('id_kasus');
             $table->string('path_file');
         });
